[refactor] EnforceDocument 을 typegoose의 DocumentType 으로 변경
#69
