+++ src/models/types/Plan.ts
@@ -1,13 +1,12 @@
 import { DocumentType } from '@typegoose/typegoose';
 import { Plan } from '@src/models/Plan';
-import { EnforceDocument } from 'mongoose';
 import { User } from '@src/models/User';
-import { UserMethods } from '@src/models/types/User';
+import { UserQueryHelpers } from '@src/models/types/User';
 
 export interface PlanMethods {
   checkPermission: (
-    this: DocumentType<Plan>,
-    user: EnforceDocument<User, UserMethods>,
+    this: DocumentType<Plan, PlanQueryHelpers>,
+    user: DocumentType<User, UserQueryHelpers>,
   ) => DocumentType<Plan>;
 }
