+++ src/models/types/User.ts
@@ -3,7 +3,9 @@ import { LoginResponse } from '@src/resolvers/types/LoginResponse';
 import { User } from '@src/models/User';
 
 export interface UserMethods {
-  getJWTToken: (this: DocumentType<User>) => Promise<LoginResponse>;
+  getJWTToken: (
+    this: DocumentType<User, UserQueryHelpers>,
+  ) => Promise<LoginResponse>;
 }
 
 export interface UserQueryHelpers {}
